UIE-190 Auth Fetch Breakout part-6  - moved appIdentifier bits out of ajax-common.ts
