+++ src/libs/ajax/leonardo/Apps.ts
@@ -4,7 +4,8 @@ import _ from 'lodash/fp';
 import * as qs from 'qs';
 import { AppAccessScope, AppToolLabel } from 'src/analysis/utils/tool-utils';
 import { authOpts } from 'src/auth/auth-session';
-import { appIdentifier, fetchLeo } from 'src/libs/ajax/ajax-common';
+import { fetchLeo } from 'src/libs/ajax/ajax-common';
+import { appIdentifier } from 'src/libs/ajax/fetch/fetch-core';
 import { CreateAppV1Request, GetAppItem, ListAppItem } from 'src/libs/ajax/leonardo/models/app-models';
 
 export const Apps = (signal: AbortSignal) => ({
