+++ src/libs/ajax/leonardo/Disks.test.ts
@@ -9,14 +9,22 @@ const mockWatchWithAuth = jest.fn();
 const mockWatchWithAppId = jest.fn();
 
 type AjaxCommonExports = typeof import('src/libs/ajax/ajax-common');
+jest.mock('src/libs/ajax/ajax-common', (): AjaxCommonExports => {
+  const mocks: AjaxCommonExports = {
+    ...jest.requireActual('src/libs/ajax/ajax-common'),
+    fetchLeo: jest.fn(),
+  };
+  return mocks;
+});
+
+type FetchCoreExports = typeof import('src/libs/ajax/fetch/fetch-core');
 type TestUtilsExports = typeof import('src/testing/test-utils');
-jest.mock('src/libs/ajax/ajax-common', (): Partial<AjaxCommonExports> => {
+jest.mock('src/libs/ajax/fetch/fetch-core', (): FetchCoreExports => {
   const { asMockedFn } = jest.requireActual<TestUtilsExports>('src/testing/test-utils');
 
-  const mocks: Partial<AjaxCommonExports> = {
-    ...jest.requireActual('src/libs/ajax/ajax-common'),
+  const mocks: FetchCoreExports = {
+    ...jest.requireActual('src/libs/ajax/fetch/fetch-core'),
     withAppIdentifier: jest.fn(),
-    fetchLeo: jest.fn(),
   };
   // mock fetch augmentors to call watcher for test assertions below.
   // (mock here so that it's baked in for module-load-time of Disks.ts)
